add xgboost layer and update layer creation process to support a platform

+++ src/create_layer.py
@@ -6,7 +6,7 @@ import zipfile
 from logging.config import fileConfig
 from pathlib import Path
 from time import sleep
-from typing import Dict, Optional
+from typing import Dict
 
 import docker
 import os
@@ -17,47 +17,39 @@ def _get_folder(folder_name: str) -> str:
     return os.path.join(Path(os.path.dirname(__file__)).parent, folder_name)
 
 
-def _get_dockerfile() -> str:
-    return os.path.join(os.path.dirname(__file__), "create_layer.Dockerfile")
+def _get_layer_full_name(layer_name: str, python_runtime: str, architecture: str) -> str:
+    return f"{layer_name}_{python_runtime.replace('.', '_')}_{architecture}"
 
 
-def get_layer_folder(layer_name: str, layer_version: str = None) -> str:
-    folder = os.path.join(_get_folder("layers"), layer_name)
-    if layer_version is not None:
-        folder = os.path.join(folder, layer_version)
-    return folder
-
-
-def get_layer_output_file(layer_name: str, layer_version: Optional[str]) -> str:
-    if layer_version is None:
-        layer_version = "latest"
-    return f"{layer_name}_{layer_version}.zip"
-
-
-def get_output_folder() -> str:
+def _get_output_folder() -> str:
     return _get_folder("outputs")
 
 
-def create_layer(layer_name: str, layer_version: Optional[str], lambda_base_image: str) -> str:
-    logging.info(f"Going to create layer: {layer_name}, Version: {layer_version}, Lambda base image: {lambda_base_image}")
-    layer_file = get_layer_output_file(layer_name, layer_version)
-    full_layer_file = os.path.join(get_output_folder(), layer_file)
+def _build_layer(layer_name: str, lambda_base_image: str, architecture: str, skip_if_exists: bool) -> str:
+    logging.info(f"Going to create layer: {layer_name}, Lambda base image: {lambda_base_image}")
+    layer_file = f"{_get_layer_full_name(layer_name, lambda_base_image, architecture)}.zip"
+    full_layer_file = os.path.join(_get_folder("outputs"), layer_file)
     if os.path.exists(full_layer_file):
-        os.remove(full_layer_file)
-    layer_folder = get_layer_folder(layer_name, layer_version)
+        if skip_if_exists:
+            logging.info("Layer file exists skipping creation")
+            return full_layer_file
+        else:
+            os.remove(full_layer_file)
+    layer_folder = os.path.join(_get_folder("layers"), layer_name)
     logging.info(f"Source files for layer found under: {layer_folder}. Going to build layer file: {layer_file}")
     client = docker.from_env()
     img, output = client.images.build(path=layer_folder, rm=True,
-                                      dockerfile=_get_dockerfile(),
+                                      dockerfile=os.path.join(os.path.dirname(__file__), "create_layer.Dockerfile"),
                                       quiet=False, nocache=False,
                                       buildargs={
                                           "LAYER_FILE": layer_file,
                                           "LAMBDA_BASE_IMAGE": lambda_base_image,
+                                          "PLATFORM": _get_platform(architecture),
                                       })
     logging.info(f"Build result:\n{''.join([str(next(iter(row.values()))) for row in output])}")
     try:
         client.containers.run(image=img.id, remove=True,
-                              mounts=[Mount(type="bind", source=get_output_folder(), target="/outputs")])
+                              mounts=[Mount(type="bind", source=_get_output_folder(), target="/outputs")])
     finally:
         sleep(3)
         img.remove()
@@ -65,7 +57,16 @@ def create_layer(layer_name: str, layer_version: Optional[str], lambda_base_imag
     return full_layer_file
 
 
-def analyze_layer_file(layer_file: str, top_n: int):
+def _get_platform(architecture: str):
+    if architecture == "arm64":
+        return "manylinux2014_aarch64"
+    elif architecture == "x86_64":
+        return "manylinux2014_x86_64"
+    else:
+        raise ValueError(f"Unknown architecture: {architecture}")
+
+
+def _analyze_layer_file(layer_file: str, top_n: int):
     with zipfile.ZipFile(layer_file) as z:
         files_to_size = {f.filename: f.compress_size for f in z.infolist()}
     dirs: Dict[str, int] = {}
@@ -90,9 +91,9 @@ def analyze_layer_file(layer_file: str, top_n: int):
     _log_dict("File types", file_types, top_n)
 
 
-def _log_dict(item_type: str, items: Dict, top_n: int) -> str:
+def _log_dict(item_type: str, items: Dict, top_n: int):
     result = f"\n========== Top {top_n} {item_type} =========="
-    for ext, size in sorted(items.items(), key=lambda x: x[1], reverse=True)[:top_n]:
+    for ext, size in sorted(items.items(), key=lambda x: x[1], reverse=True)[:min(top_n, len(items))]:
         if size > 0:
             result += f"\n{ext}, {_format_size(size)}"
     logging.info(result)
@@ -106,25 +107,31 @@ def _format_size(size: int) -> str:
     return f"{size:.2f} {unit}"
 
 
-def publish_layer(bucket: str, layer_file: str):
+def _publish_layer(layer_name: str, layer_file: str, python_runtime: str, architecture: str, bucket: str):
     import boto3
     layer_base_file_name = os.path.basename(layer_file)
     logging.info(f"Layer file name: {layer_base_file_name}, Layer size: {_format_size(os.path.getsize(layer_file))}")
 
     key = f"temp/{layer_base_file_name}"
     logging.info(f"Uploading to bucket: {bucket}. Key: {key}")
-    s3 = boto3.client("s3")
-    s3.upload_file(layer_file, bucket, key)
-
-    client = boto3.client("lambda", region_name="us-east-1")
-    response = client.publish_layer_version(
-        LayerName="sklearn",
-        Content={
-            'S3Bucket': bucket,
-            'S3Key': key
-        },
-    )
-    s3.delete_object(Bucket=bucket, Key=key)
+    session = boto3.session.Session()
+    s3_client = session.client("s3")
+    s3_client.upload_file(layer_file, bucket, key)
+    try:
+        layer_full_name = _get_layer_full_name(layer_name, python_runtime, architecture)
+        logging.info(f"Creating Layer: {layer_name}")
+        lambda_client = session.client("lambda")
+        response = lambda_client.publish_layer_version(
+            LayerName=layer_full_name,
+            Content={
+                'S3Bucket': bucket,
+                'S3Key': key
+            },
+            CompatibleRuntimes=[f"python{python_runtime}"],
+            CompatibleArchitectures=[architecture],
+        )
+    finally:
+        s3_client.delete_object(Bucket=bucket, Key=key)
     if "Version" in response:
         logging.info(f"Layer version: {response['Version']}")
     else:
@@ -134,9 +141,11 @@ def publish_layer(bucket: str, layer_file: str):
 def run(argv):
     parser = argparse.ArgumentParser()
     parser.add_argument("-l", "--layer", required=True, help="Layer name")
-    parser.add_argument("-v", "--version", required=False, help="Layer version. If no version is given use default")
-    parser.add_argument("-t", "--tag", required=False, default="latest", help="Lambda base image tag")
-    parser.add_argument("-top", "--top", required=False, default=10,
+    parser.add_argument("-r", "--runtime", required=False, default="3.12", help="Python runtime")
+    parser.add_argument("-a", "--architecture", required=False, default="x86_64", help="x86_64 or arm64")
+    parser.add_argument("-s", "--skip", required=False, default="false",
+                        help="Skip layer creation if layer already exists")
+    parser.add_argument("-t", "--top", required=False, default=10,
                         help="Top item count in layer analysis. By default top 10 items are shown")
     parser.add_argument("-p", "--publish", required=False, default="false",
                         help="Publish lambda layer")
@@ -144,14 +153,14 @@ def run(argv):
                         help="S3 bucket to use, required for the publishing process")
     args = parser.parse_args(args=argv)
 
-    layer_file = create_layer(args.layer, args.version, args.tag)
-    # layer_file = get_output_folder() + "/" + get_layer_output_file(args.layer, args.version)
-    analyze_layer_file(layer_file, args.top)
+    layer_file = _build_layer(args.layer, args.runtime, args.architecture,
+                              args.skip.lower() == "true")
+    _analyze_layer_file(layer_file, int(args.top))
     if args.publish.lower() == "true":
         if args.bucket is None:
             logging.error("Bucket must be provided for publishing")
             exit(2)
-        publish_layer(args.bucket, layer_file)
+        _publish_layer(args.layer, layer_file, args.runtime, args.architecture, args.bucket)
 
 
 if __name__ == '__main__':
